Reorganized widgets and created widget for tranformation preset application
