+++ src/components/lessons/vectors/BasisTransformations/BasisTransformations.module.css
@@ -1,0 +1,6 @@
+.widget {
+	position: absolute;
+	bottom: 10px;
+	left: 10px;
+	z-index: 2;
+}
