+++ src/components/lessons/vectors/BasisTransformations/BasisTransformations.tsx
@@ -1,0 +1,9 @@
+import styles from "./BasisTransformations.module.css";
+
+export default function BasisTransformations() {
+	return (
+		<div className={styles.widget}>
+			hello
+		</div>
+	);
+}
