+++ src/components/lessons/vectors/BasisWidget/BasisWidget.module.css
@@ -1,0 +1,64 @@
+.basis_wrapper {
+	display: flex;
+	flex-direction: row;
+	/* border: 2px solid #f4ffff; */
+}
+
+.basis_vector {
+	display: flex;
+	flex-direction: column;
+	padding-left: 2px;
+	padding-right: 2px;
+	text-align: center;
+}
+
+.basis_vector input {
+	width: 2rem;
+	text-align: center;
+	color: #f4ffff;
+	background: none;
+	border: none;
+}
+
+input::-webkit-outer-spin-button,
+input::-webkit-inner-spin-button {
+	-webkit-appearance: none;
+}
+
+.widget {
+	position: absolute;
+	top: 10px;
+	left: 10px;
+	z-index: 2;
+}
+
+.button {
+	background-color: #4b585d;
+	border: 1px solid #4b585d;
+	border-radius: 15px;
+	margin-right: 10px;
+	padding: 10px 20px;
+	cursor: pointer;
+	color: white;
+}
+
+.button:hover {
+	border: 1px solid #57ffeb;
+	transition: 0.2s ease;
+}
+
+.i {
+	color: #CD5C5C;
+}
+
+.j {
+	color: #008000;
+}
+
+.k {
+	color: #87CEEB;
+}
+
+.v {
+	color: #FFFF00;
+}
